updated form for add hbs
