+++ server.js
@@ -3,6 +3,7 @@ const path = require('path');
 const routes = require('./controllers');
 const sequelize = require('./config/connection');
 const exphbs = require('express-handlebars');
+const { extname } = require('path');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
