+++ views/add.handlebars
@@ -22,8 +22,16 @@
         <input class="form-input" type="number" id="add-year" />
       </div>
       <div class="form-group">
+        <label for="add-mileage">Mileage:</label>
+        <input class="form-input" type="number" id="add-mileage" />
+      </div>
+      <div class="form-group">
         <label for="add-condition">Condition:</label>
         <input class="form-input" type="text" id="add-condition" />
+      </div>
+      <div class="form-group">
+        <label for="add-vin">Vin Number:</label>
+        <input class="form-input" type="text" id="add-vin" />
       </div>
       <div class="form-group">
         <label for="add-link">Image Link:</label>
